tests: move some functions from algorithm tests to common.py
* _scheduled_job_entry
* _assert_distributions_equal
* make job priority optional
